requirements.txt
aws-requests-auth==0.4.3
python-dateutil==2.7.3
beautifulsoup4==4.10.0
boto3==1.17.93
justext==3.0.0
botocore==1.20.112
certifi==2022.12.07
s3transfer==0.4.2
Scrapy==2.6.3
scrapy-fake-useragent==1.4.4
scrapy-inline-requests~=0.3.1
scrapy-UserAgents==0.0.1
xmltodict==0.12.0
loguru==0.5.3
lxml==4.6.4
pyopenssl==22.0.0
cryptography==38.0.4
scrapeops-scrapy==0.5.2
scrapeops-scrapy-proxy-sdk==1.0
python-dotenv
